requirements.txt
wandb
transformers
ninja
matplotlib
numpy
PyYAML
scikit_learn
torch
tqdm
torch_geometric
torch_scatter
torch_sparse
pyg-lib